Temporary fix for bad version of cordova.js
The current npm version of cordova cli uses a wrong version of cordova.js when
doing cordova platform add <path-to-platform>. Instead of taking the file from
provided path, it downloads the published version of the platfrom and takes
cordova.js from there.

This does not happen with the latest master versions of cordova cli/lib, not
sure what fixed it. Added "cp" lines to create script to explicitly copy
cordova.js after "platform add" as a temporary workaround. Should be removed
and tested after new versions of cordova lib/cli are released.